refactor(arch): extract Flare.Infrastructure adapters (outer ring)
