+++ Flare.slnx
@@ -15,6 +15,7 @@
     <Project Path="src/Flare.Abstractions/Flare.Abstractions.csproj" />
     <Project Path="src/Flare.Components/Flare.Components.csproj" />
     <Project Path="src/Flare.Core/Flare.Core.csproj" />
+    <Project Path="src/Flare.Infrastructure/Flare.Infrastructure.csproj" />
     <Project Path="src/Flare.Theming/Flare.Theming.csproj" />
     <Project Path="src/Flare.Theme.FluentUI2/Flare.Theme.FluentUI2.csproj" />
     <Project Path="src/Flare.Theme.MaterialDesign3Expressive/Flare.Theme.MaterialDesign3Expressive.csproj" />
