+++ src/Flare.Abstractions/JsInterop/FlareJsModule.cs
