+++ src/Flare.Infrastructure/Flare.Infrastructure.csproj
@@ -1,0 +1,28 @@
+<Project Sdk="Microsoft.NET.Sdk.Razor">
+
+  <PropertyGroup>
+    <PackageId>Flare.Infrastructure</PackageId>
+    <Description>Flare browser/host adapters (outer ring): JS-interop services, theme storage, version check and the dialog/snackbar/message-box state services that implement the Flare.Abstractions ports.</Description>
+    <PackageTags>blazor;js-interop;adapters;infrastructure;dotnet</PackageTags>
+    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+    <!-- Public API must stay fully XML-documented (drives the generated API docs). -->
+    <NoWarn>$(NoWarn.Replace('CS1591',''))</NoWarn>
+    <WarningsAsErrors>$(WarningsAsErrors);CS1591</WarningsAsErrors>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <SupportedPlatform Include="browser" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Flare.Abstractions\Flare.Abstractions.csproj" />
+    <!-- The CSS-variable injector adapter uses the theming engine's token->CSS flattening (CssVarMap). -->
+    <ProjectReference Include="..\Flare.Theming\Flare.Theming.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Update="Microsoft.SourceLink.GitHub" Version="10.0.300" />
+    <PackageReference Update="MinVer" Version="7.0.0" />
+  </ItemGroup>
+
+</Project>
