+++ src/Flare.Infrastructure/JsInterop/CollisionService.cs
