+++ src/Flare.Infrastructure/JsInterop/CssVariableInjector.cs
