+++ src/Flare.Infrastructure/JsInterop/ThemeJsService.cs
