+++ tests/Flare.Components.Tests/Flare.Components.Tests.csproj
@@ -23,6 +23,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Flare.Components\Flare.Components.csproj" />
     <ProjectReference Include="..\..\src\Flare.Core\Flare.Core.csproj" />
+    <ProjectReference Include="..\..\src\Flare.Infrastructure\Flare.Infrastructure.csproj" />
   </ItemGroup>
 
   <ItemGroup>
