+++ tests/Flare.Core.Tests/Flare.Core.Tests.csproj
@@ -24,6 +24,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Flare.Core\Flare.Core.csproj" />
     <ProjectReference Include="..\..\src\Flare.Components\Flare.Components.csproj" />
+    <ProjectReference Include="..\..\src\Flare.Infrastructure\Flare.Infrastructure.csproj" />
     <!-- Reusable audit logic for the CssClasses/CSS/theme in-sync guard test. -->
     <ProjectReference Include="..\..\tools\Flare.CssAudit\Flare.CssAudit.csproj" />
   </ItemGroup>
